services/api-gateway/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
asyncpg==0.29.0
redis==5.0.1
pika==1.3.2
aio-pika==9.3.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
slowapi==0.1.9
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
structlog==24.1.0
httpx==0.25.2
websockets==12.0
